requirements.txt
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
opencv-python==4.8.0.76
Pillow==10.0.0
tqdm==4.65.0
streamlit==1.28.0
plotly==5.17.0
grad-cam==1.4.8
albumentations==1.3.1
scipy==1.11.2
requests==2.31.0
openpyxl==3.1.2
reportlab==4.0.4
fpdf2==2.7.6